(refactor) Move menu and menu-item inside shared
