--- src/app/core/core.module.ts
@@ -1,16 +1,0 @@
-import { NgModule } from '@angular/core';
-
-import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
-
-@NgModule({
-  imports: [
-  ],
-  declarations: [
-    PageNotFoundComponent
-  ],
-  exports: [
-  ],
-  providers: [
-  ]
-})
-export class CoreModule { }
