+++ src/app/intervention-survey/intervention-survey.component.html
@@ -1,6 +1,6 @@
 <md-sidenav-container>
   <md-sidenav #sidenav [mode]="mode" [align]="align">
-    <app-form-menu [items]="menuItems" (select)="select($event)" ></app-form-menu>
+    <app-menu [items]="menuItems" (select)="select($event)" ></app-menu>
   </md-sidenav>
 
   <app-toolbar-back [title]="'627 rapide-plat sud'" [mode]="mode" [sidenav]="sidenav"></app-toolbar-back>
