+++ src/app/shared/components/menu/menu.component.html
@@ -1,11 +1,11 @@
 <igo-list navigation="true"> 
   <ng-template ngFor let-item [ngForOf]="items">
-    <app-form-menu-item
+    <app-menu-item
       igoListItem
       color="accent"
       [item]=item
       (focus)="handleItemFocus(item)"
       (select)="handleItemSelect(item)">
-    </app-form-menu-item>
+    </app-menu-item>
   </ng-template>
 </igo-list>
