+++ src/app/shared/components/menu/menu.component.ts
@@ -1,10 +1,10 @@
 import { Component, Input, Output, EventEmitter } from '@angular/core';
 import { Router } from '@angular/router';
 
-import { MenuItem } from '../shared/menu-item.interface';
+import { MenuItem } from '../../models/menu-item.interface';
 
 @Component({
-  selector: 'app-form-menu',
+  selector: 'app-menu',
   templateUrl: './menu.component.html',
   styleUrls: ['./menu.component.styl']
 })
